+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Port 1.20 -> 1.20.2
+- Opening a config screen with mod menu now renders a blank background instead of the dirt background